Change the long to long long

+++ common/inc/IReadable.hpp
@@ -9,14 +9,14 @@ public:
     virtual ~IReadable() {};
 
     virtual void put(const std::string &key, const std::string &value) = 0;
-    virtual void put(const std::string &key, const unsigned long &value) = 0;
+    virtual void put(const std::string &key, const unsigned long long &value) = 0;
     virtual void put(const std::string &key, const int &value) = 0;
     virtual void put(const std::string &key, const short &value) = 0;
     virtual void put(const std::string &key, const char &value) = 0;
 
     
     virtual void get(const std::string &key, std::string &value) = 0;
-    virtual void get(const std::string &key, unsigned long &value) = 0;
+    virtual void get(const std::string &key, unsigned long long &value) = 0;
     virtual void get(const std::string &key, int &value) = 0;
     virtual void get(const std::string &key, short &value) = 0;
     virtual void get(const std::string &key, char &value) = 0;
